+++ src/lib.rs
@@ -105,6 +105,8 @@ pub mod delay;
 pub mod gpio;
 #[cfg(feature = "device-selected")]
 pub mod i2c;
+#[cfg(feature = "i2s")]
+pub mod i2s;
 #[cfg(all(
     feature = "usb_fs",
     any(